fixed a bug
I have fixed a bug where when you flag a square, then click on it
normally, the flaged attribute of the square does not change, so the
game thinks there are more flagged squares then there actually are.
I fixed this by not allowing you to click a square unless it is not flagged.